fix: Convert USD cents to satoshis for proper payment amounts
- Change settlementAmount from long to decimal to preserve fractional values
- Add USD wallet detection to transaction processing
- Convert USD cents to satoshis using exchange rate for USD wallets
- Fix invoice payment amounts showing 1 sat instead of full amount

The Flash API returns settlement amounts in USD cents for USD wallets,
not satoshis. This fix properly converts these values to satoshis
for BTCPay Server compatibility.

+++ Services/BoltcardInvoicePoller.cs
@@ -129,7 +129,7 @@ namespace BTCPayServer.Plugins.Flash.Services
                             // Mark as paid
                             await invoiceService.MarkInvoiceAsPaidAsync(
                                 matchingInvoice.Key, 
-                                Math.Abs(payment.SettlementAmount ?? 0));
+                                (long)Math.Abs(payment.SettlementAmount ?? 0));
                         }
                         else if (payment.Memo?.Contains("Boltcard", StringComparison.OrdinalIgnoreCase) == true ||
                                  Math.Abs(payment.SettlementAmount ?? 0) < 1000)
@@ -139,7 +139,7 @@ namespace BTCPayServer.Plugins.Flash.Services
                             _logger.LogInformation("[BOLTCARD POLLER] Marking likely Boltcard payment as paid: {Id}", payment.Id);
                             await invoiceService.MarkInvoiceAsPaidAsync(
                                 payment.Id, 
-                                Math.Abs(payment.SettlementAmount ?? 0));
+                                (long)Math.Abs(payment.SettlementAmount ?? 0));
                         }
                     }
                 }
